requirements.txt
django==4.2.11
django-cors-headers==4.4.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-redis==5.4.0
python-dotenv==1.0.1
requests==2.32.3
djangorestframework-simplejwt==5.4.0
pyjwt==2.8.0
drf-yasg==1.21.7
redis==5.0.0b2